% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad2_stat.R
\name{ad2_stat}
\alias{ad2_stat}
\title{Computes Two-sample Anderson-Darling statistic}
\usage{
ad2_stat(x, y)
}
\arguments{
\item{x}{Numeric vector (first sample).}

\item{y}{Numeric vector (second sample).}
}
\value{
Two-sample Anderson-Darling statistic.
}
\description{
Computes Two-sample Anderson-Darling statistic
}
\examples{
x <- rnorm(100)
y <- rnorm(100, mean = 0.5)
ad2_stat(x, y)
}
