% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_initial_abs_increase.R
\name{check_initial_abs_increase}
\alias{check_initial_abs_increase}
\title{Check if Absolute Values Initially Increase}
\usage{
check_initial_abs_increase(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
1 if absolute values start increasing, 0 otherwise.
}
\description{
Check if Absolute Values Initially Increase
}
\examples{
check_initial_abs_increase(c(-1, -2, -3, -2, -1))  # Returns 0
check_initial_abs_increase(c(0.1, 0.5, 1.0, 2.0, 3.0))  # Returns 1
}
