% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sign_change.R
\name{check_sign_change}
\alias{check_sign_change}
\title{Find the  Sign Change in a Series}
\usage{
check_sign_change(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The index of the first sign change, or NA if no sign change occurs.
}
\description{
Find the  Sign Change in a Series
}
\examples{
check_sign_change(c(-1, -2, -3, 3, 4))  # Returns 4
check_sign_change(c(5, 4, 3, 2, 1))  # Returns NA (no sign change)
check_sign_change(c(0, 0, 1, -1, 2))  # Returns 4
}
