% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_first_sign_change.R
\name{find_first_sign_change}
\alias{find_first_sign_change}
\title{Find the First Sign Change in a Series}
\usage{
find_first_sign_change(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The index of the first sign change, or NA if no sign change occurs.
}
\description{
Find the First Sign Change in a Series
}
\examples{
find_first_sign_change(c(-1, -2, -3, 3, 4))  # Returns 4
find_first_sign_change(c(5, 4, 3, 2, 1))  # Returns NA (no sign change)
#'
}
