% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siv_regression.R
\name{siv_regression}
\alias{siv_regression}
\title{Perform Instrumental Variable Regression using SIV}
\usage{
siv_regression(data, Y, X, H, reps)
}
\arguments{
\item{data}{A data frame.}

\item{Y}{Name of the dependent variable.}

\item{X}{Name of the endogenous variable.}

\item{H}{Vector of exogenous variable names.}

\item{reps}{Number of bootstrap loops.}
}
\value{
A list containing OLS and IV regression results.
}
\description{
Perform Instrumental Variable Regression using SIV
}
\examples{
df <- wooldridge::mroz  # Use sample data set
data <- df[complete.cases(df), ]  # Remove missing values
result <- siv_regression(data, "hours", "lwage", c("educ", "age", "kidslt6", "kidsge6", "nwifeinc"), reps=5)
summary(result$IV2, diagnostics=TRUE)
}
