% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siv_regression.R
\name{siv_regression}
\alias{siv_regression}
\title{Perform Instrumental Variable Regression using SIV}
\usage{
siv_regression(data, Y, X, H, reps)
}
\arguments{
\item{data}{A data frame.}

\item{Y}{Name of the dependent variable.}

\item{X}{Name of the endogenous variable.}

\item{H}{Vector of exogenous variable names.}

\item{reps}{Number of bootstrap loops.}
}
\value{
A list containing SIV regression results.IV1- a simple SIV, IV2- a parametric SIV, IV3- a non-parametric SIV, citable- a table of confidence intervals.
}
\description{
Perform Instrumental Variable Regression using SIV
}
\examples{
df <- wooldridge::mroz  # Use sample data set
data <- df[complete.cases(df), ]  # Remove missing values
attach(data)
H0 <- data.frame(hours, lwage, educ, age, kidslt6, kidsge6, nwifeinc)
Y <- as.character(colnames(H0))[1] ###OUTCOME variable
X <-as.character(colnames(H0))[2] ###ENDOEGENOUS variable
H<- as.character(colnames(H0))[-(1:2)] #
result <- siv_regression(data, Y, X, H, reps=5)
iv1 <-(result$IV1)
iv2 <-(result$IV2)
iv3 <-(result$IV3)
summ.iv1 <- summary(iv1, diagnostics=TRUE)
summ.iv2<- summary(iv2, diagnostics=TRUE)
summ.iv3 <- summary(iv3, diagnostics=TRUE)
result$citable ###  renders the CI table of beta estimates
}
